Use default arrays instead of uninit arrays
In the rasterizer, arrays of data are gradually initialized by creating
an uninitialized array and writing into those arrays. However, this is
technically unsound in Rust, and causes MIRI to fail when I try to run
on it.

Out of curiousity, I tried switching it to zeroing the array first
before writing to it. However, by some benchmarks I wrote it seems to
have no effect on the speed of the system. So I think it's a viable
solution to reduce undefined behavior.

Signed-off-by: John Nunley <dev@notgull.net>